id0nknow

+++ lib/screens/screen_time.dart
@@ -56,12 +56,12 @@ class _TimeScreenState extends State<TimeScreen> {
 
   void graphData() {
     apps?.forEach((key, app) {
-      childAppsDataDaily.add(
-          _ChartData(app['app_name'], app['current_day_screen_time'] ?? 0));
-      childAppsDataMonthly.add(
-          _ChartData(app['app_name'], app['current_month_screen_time'] ?? 0));
-      childAppsDataWeekly.add(
-          _ChartData(app['app_name'], app['current_week_screen_time'] ?? 0));
+      childAppsDataDaily.add(_ChartData(
+          app['app_name'], app['current_day_screen_time'] ?? 0));
+      childAppsDataMonthly.add(_ChartData(
+          app['app_name'], app['current_month_screen_time'] ?? 0));
+      childAppsDataWeekly.add(_ChartData(
+          app['app_name'], app['current_week_screen_time'] ?? 0));
 
       totalAppHrsDaily =
           totalAppHrsDaily + (app['current_day_screen_time'] ?? 0);
@@ -251,5 +251,5 @@ class _ChartData {
   _ChartData(this.x, this.y);
 
   final String x;
-  final double y;
+  final int y;
 }
